requirements.txt
asgiref==3.3.4
Django==3.2.4
django-cleanup==5.2.0
django-cors-headers==3.7.0
django-crispy-forms==1.12.0
django-decouple==2.1
djangorestframework==3.12.4
drf-nested-serializer==0.1.0
filetype==1.0.7
gunicorn==20.1.0
Pillow==8.2.0
psycopg2-binary==2.9.1
pytz==2021.1
sqlparse==0.4.1
